tatartwin/requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.3.0
captcha==0.3
celery==4.4.5
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
DAWG-Python==0.7.2
defusedxml==0.7.0rc1
Django==3.0.7
django-appconf==1.0.4
django-celery==3.3.1
django-celery-email==3.0.0
django-cors-headers==3.3.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-rest-framework==0.1.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
docopt==0.6.2
future==0.18.2
idna==2.9
kombu==4.6.10
oauthlib==3.1.0
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-memcached==1.59
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
